requirements.txt
pandas
pyyaml
networkx
matplotlib
streamlit
pyvis